Add python function to convert truth data from dataframe to FrameContainer
A new convenience function to convert truth data from a pandas dataframe
to the Cap'n Proto FrameContainer was added. You basically just have to
change the column names to the corresponding attributes of the Cap'n
Proto Schema and this function will do the hard work for you.

Also adds pandas as new dependency of this project. This also is just
for convenience and could be refactored if necessary.